Implement get_physical() by parsing /proc/cpuinfo
In a lot of scenarios what you want to know is the number of physical
cores and not the logical ones. For a lot of algorithms hyperthreading
is not a performance boost or only a minimal one. Unfortunately it
seems Linux doesn't have a way to fetch that information easily. So
parse /proc/cpuinfo and fallback on get() if that fails at any point.