Use Vec instead of HashMap when counting physical cores
The current implementation uses a HashMap to deduplicate the output
from each core of the same cpu.  This commit instead collects the
output for each core in to a Vec, and then sorts it to deduplicate
physical CPUs.

This reduces the code size processed by LLVM by 15-20%, as counted
by `cargo llvm-lines --lib -p num_cpus` on both debug and release.

These implementations have different performance characteristics:
- the HashMap must hash each key, and SipHash is slow on small keys
- the number of cores will be small (<1024) so sorting the list
  should be very fast
- the list will likely already be sorted

I have not benchmarked this code, but it should be around the same
speed or slightly faster (from testing against randomized lists).